Initialized PIXI/webgl
