+++ src/foundry/client/pixi/webgl/extensions/blend-modes.d.ts
