+++ src/foundry/client/pixi/webgl/extensions/index.d.ts
@@ -1,0 +1,3 @@
+import "./batch-renderer";
+import "./batch-shader-generator";
+import "./blend-modes";
