+++ src/foundry/client/pixi/webgl/helpers/index.d.ts
@@ -1,0 +1,3 @@
+import "./framebugger-snapshot";
+import "./smooth-noise";
+import "./texture-extractor";
