+++ src/foundry/client/pixi/webgl/shaders/effects/index.d.ts
@@ -1,0 +1,3 @@
+import "./lighting";
+import "./vision";
+import "./weather";
