+++ src/foundry/client/pixi/webgl/shaders/effects/lighting.d.ts
