+++ src/foundry/client/pixi/webgl/shaders/effects/weather/index.d.ts
@@ -1,0 +1,5 @@
+import "./base";
+import "./effect";
+import "./fog";
+import "./rain";
+import "./snow";
