+++ src/foundry/client/pixi/webgl/shaders/filters/index.d.ts
@@ -1,0 +1,3 @@
+import "./blur";
+import "./filters";
+import "./fxaa";
