+++ src/foundry/client/pixi/webgl/shaders/index.d.ts
@@ -1,0 +1,4 @@
+import "./effects";
+import "./filters";
+import "./samplers";
+import "./base";
