+++ src/foundry/client/pixi/webgl/shaders/samplers/index.d.ts
@@ -1,0 +1,6 @@
+import "./color-adjustment";
+import "./fog-of-war";
+import "./inverse-occlusion";
+import "./invisibility";
+import "./monochromatic";
+import "./occlusion";
